+++ specfunc/gamma.v
@@ -14,7 +14,7 @@ import vsl.vmath
 // masks any imprecision in the polynomial.
 fn stirling(x f64) (f64, f64) {
 	if x > 200 {
-		return vmath.inf(1), f64(1.0)
+		return vmath.inf(1), 1.0
 	}
 	sqrt_two_pi := 2.506628274631000502417
 	max_stirling := 143.01608
